refactor: extract shared-services module and move StateModels to taskrun-domain (Phase 0a/0b)
Extract FileService, HttpAIClient, StateService, and RateLimiter from
app.control into a new shared-services module (shared.services package).
This unblocks moving domain control layers into their modules since they
can now depend on shared-services instead of root.

Also move StateModels.scala from root taskrun/entity to taskrun-domain
module, replacing shared.entity.GatewayConfig alias with direct
config.entity.GatewayConfig reference. Remove the now-empty
shared/entity/CommonTypes.scala.
