+++ modules/shared-services/src/main/scala/shared/services/FileService.scala
@@ -1,4 +1,4 @@
-package app.control
+package shared.services
 
 import java.io.IOException
 import java.nio.charset.StandardCharsets
