+++ modules/shared-services/src/main/scala/shared/services/HttpAIClient.scala
@@ -1,4 +1,4 @@
-package app.control
+package shared.services
 
 import zio.*
 import zio.http.*
