+++ modules/shared-services/src/main/scala/shared/services/RateLimiter.scala
@@ -1,4 +1,4 @@
-package app.control
+package shared.services
 
 import zio.*
