+++ modules/shared-services/src/main/scala/shared/services/StateService.scala
@@ -1,4 +1,4 @@
-package app.control
+package shared.services
 
 import java.nio.file.Path
 import java.security.MessageDigest
