+++ modules/taskrun-domain/src/main/scala/taskrun/entity/StateModels.scala
@@ -1,18 +1,19 @@
 package taskrun.entity
 
-import java.nio.file.{ Path, Paths }
 import java.time.Instant
 
 import zio.json.*
 
+import shared.json.JsonCodecs.given
+
 /** Metadata for workspace associated with a task run */
 case class WorkspaceMetadata(
   runId: String,
-  workspaceRoot: Path,
-  stateDir: Path,
-  reportsDir: Path,
-  outputDir: Path,
-  tempDir: Path,
+  workspaceRoot: java.nio.file.Path,
+  stateDir: java.nio.file.Path,
+  reportsDir: java.nio.file.Path,
+  outputDir: java.nio.file.Path,
+  tempDir: java.nio.file.Path,
   createdAt: Instant,
 ) derives JsonCodec
 
@@ -32,7 +33,7 @@ case class TaskState(
   completedSteps: Set[TaskStep],
   artifacts: Map[String, String],
   errors: List[TaskError],
-  config: shared.entity.GatewayConfig,
+  config: _root_.config.entity.GatewayConfig,
   workspace: Option[WorkspaceMetadata],
   status: TaskStatus,
   lastCheckpoint: Instant,
@@ -44,7 +45,7 @@ case class Checkpoint(
   runId: String,
   step: String,
   createdAt: Instant,
-  artifactPaths: Map[String, Path],
+  artifactPaths: Map[String, java.nio.file.Path],
   checksum: String,
 ) derives JsonCodec
 
@@ -64,13 +65,3 @@ case class TaskRunSummary(
   startedAt: Instant = Instant.EPOCH,
   updatedAt: Instant = Instant.EPOCH,
 ) derives JsonCodec
-
-given JsonCodec[Path] = JsonCodec[String].transform(
-  str => Paths.get(str),
-  path => path.toString,
-)
-
-given JsonCodec[Instant] = JsonCodec[String].transform(
-  str => Instant.parse(str),
-  instant => instant.toString,
-)
