+++ src/it/scala/integration/AnalysisPipelineIntegrationSpec.scala
@@ -9,7 +9,7 @@ import zio.test.*
 import agent.entity.{ Agent, AgentEvent, AgentRepository }
 import analysis.control.{ AnalysisAgentRunner, AnalysisAgentRunnerLive }
 import analysis.entity.{ AnalysisDoc, AnalysisEvent, AnalysisRepository, AnalysisType }
-import app.control.FileService
+import shared.services.FileService
 import _root_.config.entity.SettingRow
 import db.TaskRepository
 import taskrun.entity.{ TaskArtifactRow, TaskReportRow, TaskRunRow }
