+++ src/main/scala/analysis/control/AnalysisAgentRunner.scala
@@ -11,7 +11,7 @@ import _root_.config.entity.{ AIProvider, AIProviderConfig }
 import agent.control.BuiltInAgentSynchronizer
 import agent.entity.{ Agent, AgentRepository }
 import analysis.entity.{ AnalysisDoc, AnalysisEvent, AnalysisRepository, AnalysisType }
-import app.control.FileService
+import shared.services.FileService
 import db.TaskRepository
 import llm4zio.core.*
 import llm4zio.providers.{ GeminiCliExecutionContext, GeminiCliExecutor, HttpClient }
