--- src/main/scala/shared/entity/CommonTypes.scala
@@ -1,4 +1,0 @@
-package shared.entity
-
-type GatewayConfig = _root_.config.entity.GatewayConfig
-val GatewayConfig = _root_.config.entity.GatewayConfig
