+++ src/test/scala/checkpoint/control/CheckpointReviewServiceSpec.scala
@@ -7,7 +7,7 @@ import zio.*
 import zio.test.*
 
 import _root_.config.entity.{ MigrationConfig, WorkflowDefinition }
-import app.control.StateService
+import shared.services.StateService
 import checkpoint.entity.*
 import conversation.entity.api.{ ChatConversation, ConversationEntry, MessageType, SenderType, SessionContextLink }
 import db.ChatRepository
